Assign 5 0.0004 : I couldn't figure it out
In theory this should be working, and I have legitimately no idea why it keeps crashing on load. If you want a demonstration as to what I am legitimatly capable of doing when it comes to Lighting, please please please, check out Project_1. It's configured, it has lighting, it has a 3D model, and it works. I don't know why this isn't working.

Okay, have a good night
-Bella